@@ -11,20 +11,25 @@
             "*": "dist"
         }
     },
-    "require": {
-        "php": ">=7.2",
-        "simplesamlphp/composer-module-installer": "~1.1",
-        "webmozart/assert": "1.4"
+    "autoload": {
+        "psr-4": {
+            "Simplesamlphp\\Module\\casserver\\": "lib"
+        }
     },
     "autoload-dev": {
         "psr-4": {
-            "SimpleSAML\\Test\\": "vendor/simplesamlphp/simplesamlphp/tests"
+            "Simplesamlphp\\Casserver\\": "tests/lib"
         }
     },
+    "require": {
+        "php": ">=7.4 || ^8.0",
+
+        "simplesamlphp/composer-module-installer": "~1.1",
+        "simplesamlphp/simplesamlphp": "^2.0.0-beta.1",
+        "simplesamlphp/xml-common": "^0.8.1"
+    },
     "require-dev": {
-        "simplesamlphp/simplesamlphp": "^1.18",
-        "simplesamlphp/simplesamlphp-test-framework": "~0.1.0",
-        "squizlabs/php_codesniffer": "^3.5"
+        "simplesamlphp/simplesamlphp-test-framework": "~1.1.4"
     },
     "support": {
         "issues": "https://github.com/simplesamlphp/simplesamlphp-module-casserver/issues",
