@@ -21,6 +21,8 @@
  *
  */
 
+declare(strict_types=1);
+
 namespace SimpleSAML\Module\casserver\Cas\Ticket;
 
 use SimpleSAML\Configuration;
@@ -29,10 +31,10 @@ use SimpleSAML\Store\Redis;
 class RedisTicketStore extends TicketStore
 {
     /** @var string $prefix */
-    private $prefix = '';
+    private string $prefix = '';
 
     /** @var \SimpleSAML\Store\Redis $redis */
-    private $redis;
+    private Redis $redis;
 
 
     /**
@@ -56,7 +58,7 @@ class RedisTicketStore extends TicketStore
      * @param string $ticketId
      * @return array|null
      */
-    public function getTicket($ticketId)
+    public function getTicket(string $ticketId): ?array
     {
         return $this->redis->get($this->prefix, $ticketId);
     }
@@ -64,9 +66,8 @@ class RedisTicketStore extends TicketStore
 
     /**
      * @param array $ticket
-     * @return void
      */
-    public function addTicket(array $ticket)
+    public function addTicket(array $ticket): void
     {
         $this->redis->set($this->prefix, $ticket['id'], $ticket, $ticket['validBefore']);
     }
@@ -74,9 +75,8 @@ class RedisTicketStore extends TicketStore
 
     /**
      * @param string $ticketId
-     * @return void
      */
-    public function deleteTicket($ticketId)
+    public function deleteTicket(string $ticketId): void
     {
         $this->redis->delete($this->prefix, $ticketId);
     }
