@@ -1,9 +1,11 @@
 <?php
 
+declare(strict_types=1);
+
 namespace SimpleSAML\Module\casserver\Cas\Protocol;
 
 use SimpleSAML\Configuration;
-use SimpleSAML\XML\Shib13\AuthnResponse;
+use SimpleSAML\Module\casserver\Shib13\AuthnResponse;
 
 class SamlValidateResponder
 {
@@ -13,7 +15,7 @@ class SamlValidateResponder
      * @param array $ticket The cas ticket
      * @return string The saml 1 xml for the CAS response
      */
-    public function convertToSaml(array $ticket)
+    public function convertToSaml(array $ticket): string
     {
         $serviceUrl = $ticket['service'];
         $attributes = $ticket['attributes'];
@@ -57,7 +59,7 @@ class SamlValidateResponder
      * @param string $samlResponse
      * @return string
      */
-    public function wrapInSoap($samlResponse)
+    public function wrapInSoap(string $samlResponse): string
     {
         $envelope = <<<SOAP
 <?xml version="1.0" encoding="utf-8"?>
