@@ -1,6 +1,6 @@
 <?php
 
-/*
+/**
  * Configuration for the module casserver.
  *
  */
