@@ -1,17 +1,19 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Simplesamlphp\Casserver;
 
+use PHPUnit\Framework\TestCase;
 use SimpleSAML\Configuration;
 use SimpleSAML\Module\casserver\Cas\AttributeExtractor;
 
-class AttributeExtractorTest extends \PHPUnit\Framework\TestCase
+class AttributeExtractorTest extends TestCase
 {
     /**
      * Confirm behavior of a default configuration
-     * @return void
      */
-    public function testNoCasConfig()
+    public function testNoCasConfig(): void
     {
         $casConfig = [
             // Default is to use eppn and copy all attributes
@@ -34,9 +36,8 @@ class AttributeExtractorTest extends \PHPUnit\Framework\TestCase
 
     /**
      * Test disable attribute copying
-     * @return void
      */
-    public function testNoAttributeCopying()
+    public function testNoAttributeCopying(): void
     {
         $casConfig = [
             'attributes' => false
@@ -59,9 +60,8 @@ class AttributeExtractorTest extends \PHPUnit\Framework\TestCase
 
     /**
      * Confirm customizing the attribute for user and attributes to copy
-     * @return void
      */
-    public function testCustomAttributeCopy()
+    public function testCustomAttributeCopy(): void
     {
         $casConfig = [
             'attrname' => 'userNameAttribute',
@@ -88,9 +88,8 @@ class AttributeExtractorTest extends \PHPUnit\Framework\TestCase
 
     /**
      * Confirm empty authproc has no affect
-     * @return void
      */
-    public function testEmptyAuthproc()
+    public function testEmptyAuthproc(): void
     {
         $casConfig = [
             // Default is to use eppn and copy all attributes
@@ -114,9 +113,8 @@ class AttributeExtractorTest extends \PHPUnit\Framework\TestCase
 
     /**
      * Test authproc configurations can adjust the attributes.
-     * @return void
      */
-    public function testAuthprocConfig()
+    public function testAuthprocConfig(): void
     {
         // Authproc filters need a config.php defined
         putenv('SIMPLESAMLPHP_CONFIG_DIR=' . dirname(__DIR__) . '/config/');
