@@ -1,16 +1,18 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Simplesamlphp\Casserver;
 
+use PHPUnit\Framework\TestCase;
 use SimpleSAML\Module\casserver\Cas\Protocol\SamlValidateResponder;
 use SimpleXMLElement;
 
-class SamlValidateTest extends \PHPUnit\Framework\TestCase
+class SamlValidateTest extends TestCase
 {
     /**
-     * @return void
      */
-    public function testSamlValidatXmlGeneration()
+    public function testSamlValidatXmlGeneration(): void
     {
         $serviceUrl = 'http://jellyfish.greatvalleyu.com:7777/ssomanager/c/SSB';
         $udcValue = '2F10C881AC7D55942329E149405DC2F5';
