@@ -25,6 +25,6 @@ session_cache_limiter('nocache');
 
 $globalConfig = \SimpleSAML\Configuration::getInstance();
 
-$t = new \SimpleSAML\XHTML\Template($globalConfig, 'casserver:loggedIn.php');
+$t = new \SimpleSAML\XHTML\Template($globalConfig, 'casserver:loggedIn.twig');
 
-$t->show();
+$t->send();
